Changed a bit how witness work.
Now witness can be of two types: ActualWitness, which hold a theorem that is exactly the goal, and ImplicationWitness, which contains a theorem which should be used for deriving the actual Goal. Both have their advantages and disadvantages. The actual version can be checked for success by the user directly when a tries to compute a Witness, while the other one offers a slightly less verbose syntax.